Add jsdoc parsing helpers, gather `docurl` parsing togeher
