--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/comments/getNodeJsDocTags.ts
@@ -1,5 +1,0 @@
-import * as ts from 'typescript';
-
-export function getNodeJsDocTags(node: ts.Node): ts.JSDocTag[] {
-  return Array.from(ts.getJSDocTags(node));
-}
