+++ packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/comments/getUXPinNamespaceComment.ts
@@ -1,6 +1,6 @@
 import * as ts from 'typescript';
 import { CommentTags } from '../../../CommentTags';
-import { getNodeJsDocTag } from './getNodeJsDocTag';
+import { getNodeJsDocTag } from './jsdoc-helpers';
 
 export function getUXpinNamespaceComment(node: ts.Node): ts.JSDocTag | undefined {
   return getNodeJsDocTag(node, CommentTags.UXPIN_NAMESPACE);
