--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/comments/getUXpinDocUrlComment.ts
@@ -1,7 +1,0 @@
-import * as ts from 'typescript';
-import { CommentTags } from '../../../CommentTags';
-import { getNodeJsDocTag } from './getNodeJsDocTag';
-
-export function getUXpinDocUrlComment(node: ts.Node): ts.JSDocTag | undefined {
-  return getNodeJsDocTag(node, CommentTags.UXPIN_DOC_URL);
-}
