+++ packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/comments/hasUXPinComponentComment.ts
@@ -1,6 +1,6 @@
 import * as ts from 'typescript';
 import { CommentTags } from '../../../CommentTags';
-import { getNodeJsDocTag } from './getNodeJsDocTag';
+import { getNodeJsDocTag } from './jsdoc-helpers';
 
 export function hasUXPinComponentComment(node: ts.Node): boolean {
   return getNodeJsDocTag(node, CommentTags.UXPIN_COMPONENT) !== undefined;
