+++ packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/comments/jsdoc-helpers.ts
@@ -1,10 +1,12 @@
 import * as ts from 'typescript';
-import { CommentTags } from '../../../CommentTags';
-import { getNodeJsDocTags } from './getNodeJsDocTags';
+
+export function getNodeJsDocAllTags(node: ts.Node): ts.JSDocTag[] {
+  return Array.from(ts.getJSDocTags(node));
+}
 
 export function getNodeJsDocTag(node: ts.Node, tag: CommentTags): ts.JSDocTag | undefined {
   const nodeWithJsDoc = getNodeWithJsDoc(node);
-  const comments: ts.JSDocTag[] = getNodeJsDocTags(nodeWithJsDoc);
+  const comments: ts.JSDocTag[] = getNodeJsDocAllTags(nodeWithJsDoc);
 
   return comments.find((comment) => comment.getText().trim().includes(tag));
 }
