--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/component/getComponentDocUrl.ts
@@ -1,8 +1,0 @@
-import * as ts from 'typescript';
-import { getUXpinDocUrlComment } from '../comments/getUXpinDocUrlComment';
-
-export function getComponentDocUrl(component: ts.Node): string | undefined {
-  const componentDocUrl: ts.JSDocTag | undefined = getUXpinDocUrlComment(component);
-
-  return componentDocUrl?.comment;
-}
