+++ packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/component/getComponentWrappers.ts
@@ -2,7 +2,7 @@ import * as ts from 'typescript';
 import { CommentTags } from '../../../CommentTags';
 import { ComponentWrapper } from '../../../wrappers/ComponentWrapper';
 import { parseWrapperAnnotation } from '../../../wrappers/parseWrapperAnnotation';
-import { getNodeJsDocTag } from '../comments/getNodeJsDocTag';
+import { getNodeJsDocTag } from '../comments/jsdoc-helpers';
 import { ComponentDeclaration } from './getPropsTypeAndDefaultProps';
 
 export function getComponentWrappers(component: ComponentDeclaration): ComponentWrapper[] {
